apps/categories/serializers/category.py
from rest_framework import serializers

from apps.categories.models import Category


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ["id", "name", "description"]
        extra_kwargs = {
            "id": {"read_only": True},
        }
