requirements.txt
amqp==5.3.1
asgiref==3.9.2
attrs==25.4.0
billiard==4.2.2
boto3==1.42.3
cachetools==6.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
channels==4.3.2
channels-redis==4.3.0
msgpack==1.1.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
distlib==0.4.0
daphne==4.2.1
Django==5.2.9
django-celery-beat==2.8.1
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-filter==25.2
django-js-asset==3.1.2
django-redis==6.0.0
django-silk==5.4.3
django-storages==1.14.6
django-stubs==5.2.5
django-stubs-ext==5.2.5
django-timezone-field==7.1
django-unfold==0.71.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-nested-routers==0.95.0
filelock==3.19.1
gevent==25.5.1
google-auth==2.43.0
gprof2dot==2025.4.14
greenlet==3.2.4
gunicorn==23.0.0
identify==2.6.15
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
nodeenv==1.9.1
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.11.0
zope.event==6.0
zope.interface==8.0.1
